Better validation if data was scrapped correctly, increased timeout for fxtwitter and ddinstagram. Not deleting media if it is recent (so we dont run in to being used errors)

+++ Services/ClearTempBackgroundService.cs
@@ -27,8 +27,11 @@ public class ClearTempBackgroundService : BackgroundService
             {
                 try
                 {
-                    file.Delete();
-                    _logger.LogInformation("Delted {file}", file.FullName);
+                    if (file.CreationTimeUtc < DateTime.UtcNow.AddMinutes(-5))  //not delete because if it recent could be in use
+                    {
+                        file.Delete();
+                        _logger.LogInformation("Delted {file}", file.FullName);
+                    }
                 }
                 catch
                 {
